rename file, update toolbar, move coord editors
