+++ web/client/components/mapcontrols/search/SearchBar.jsx
@@ -11,7 +11,7 @@ const React = require('react');
 const {FormControl, FormGroup, Glyphicon, Row, Col} = require('react-bootstrap');
 const LocaleUtils = require('../../../utils/LocaleUtils');
 const DropdownToolbarOptions = require('../../misc/toolbar/DropdownToolbarOptions');
-const CoordinateEntry = require('../../mapcontrols/annotations/CoordinateEntry');
+const CoordinateEntry = require('../../misc/coordinateeditors/CoordinateEntry');
 const Toolbar = require('../../misc/toolbar/Toolbar');
 const Spinner = require('react-spinkit');
 const {isNumber} = require('lodash');
