+++ web/client/components/misc/coordinateeditors/CoordinateEntry.jsx
