+++ web/client/components/misc/coordinateeditors/editors/DecimalCoordinateEditor.jsx
