+++ web/client/components/misc/coordinateeditors/enhancers/coordinateTypePreset.js
