+++ web/client/components/misc/coordinateeditors/enhancers/decimalToAeronautical.js
