+++ web/client/epics/map.js
@@ -16,7 +16,7 @@ const {isSupportedLayer} = require('../utils/LayersUtils');
 const {warning} = require('../actions/notifications');
 const {resetControls} = require('../actions/controls');
 const {clearLayers} = require('../actions/layers');
-const {removeAllAdditionalLayer} = require('../actions/additionallayers');
+const {removeAllAdditionalLayers} = require('../actions/additionallayers');
 const {head} = require('lodash');
 
 const handleCreationBackgroundError = (action$, store) =>
@@ -69,7 +69,7 @@ const handleCreationLayerError = (action$, store) =>
     });
 
 const resetMapOnInit = action$ =>
-    action$.ofType(INIT_MAP).switchMap(() => Rx.Observable.of(removeAllAdditionalLayer(), resetControls(), clearLayers()));
+    action$.ofType(INIT_MAP).switchMap(() => Rx.Observable.of(removeAllAdditionalLayers(), resetControls(), clearLayers()));
 
 module.exports = {
     handleCreationLayerError,
