+++ web/client/reducers/additionallayers.js
@@ -11,7 +11,7 @@ const {
     UPDATE_ADDITIONAL_LAYER,
     REMOVE_ADDITIONAL_LAYER,
     UPDATE_OPTIONS_BY_OWNER,
-    REMOVE_ALL_ADDITIONAL_LAYER
+    REMOVE_ALL_ADDITIONAL_LAYERS
 } = require('../actions/additionallayers');
 const { head, pickBy, identity, isObject, isArray } = require('lodash');
 
@@ -48,7 +48,7 @@ function additionallayers(state = [], action) {
             const {id, owner} = action;
             return owner ? state.filter(layerItem => layerItem.owner !== owner) : state.filter(layerItem => layerItem.id !== id);
         }
-        case REMOVE_ALL_ADDITIONAL_LAYER: {
+        case REMOVE_ALL_ADDITIONAL_LAYERS: {
             return [];
         }
         default:
